ril: fix robotic-voice issue
RIL_UNSOL_WB_AMR_STATE is not supported anymore by stock LP or MM ril.
So the callback solution to be informed about the WB-AMR state does not
work. We have to check for an AM event with the
com.samsung.intent.action.WB_AMR instead. This is done by parsing all AM
events.

An alternative solution would have been to register for the intent in an
additional Java app.

Change-Id: I2f20bfa6a4cadad6a3beccd2232d2bd38f8ad0e4