ril: fix error when SMS send
when an SMS is send an error message is displayed in the messenger app
that the SMS could not be send. Contrary to error message the SMS is
normally send.

In the stock MM ril the RIL_SMS_Response type was extended by a new
4-byte field. Because of this the response type size check fails.

In addition RIL_UNSOL_STK_SEND_SMS_RESULT should be handled as it leads
to an exception.

Change-Id: Ie5d7225bdce1d676b74b7f94d88ad4fde132b9d9