patcher: link libsec-ril.so to Google Protobuffer 2.6 compatibilty lib.
Oreo uses Google Protobuf 3.0 which is incompatible with libsec-ril.
As gpsd links libsec-ril.so and libhwui.so (which needs Protobuf 3.0) we
have to link Protobuf 2.6 as well as 3.0.

The compatibility lib for 2.6 (libprotobuf-cpp-fl26.so) is provided by
the cm-3470/android_external_protobuf-compat-2.6 repo.
To avoid namespace clashes the namespace was renamed from "google::" to
"gxxgle::".

We also have to rename the namespaces in libsec-ril.so to be able to
link against libprotobuf-cpp-fl26.so.

Change-Id: I51e0acd20a7e9bda413dd85d762581f1215a255e