allocator: add support for passthrough mode
With AOSP only hwbinder can be used to find an allocator, passthrough
mode is not supported.

This patch generates a library
/system/lib/hw/android.hidl.allocator@1.0-impl.so besides the hw binder
service /system/bin/hw/android.hidl.allocator@1.0-service. The shared
library implements HIDL_FETCH_IAllocator so that it can be used for
passthrough mode. /system/manifest.xml must be changed to passthrough
mode, otherwise the hwbinder service is used.

Change-Id: I66e28cab0d765fd24f84001db096f68006d0e189